Camada Controls: Adicionado a classe RegisterProductMethods, para que a mesma receba os métodos de seu respectivo form.
